Modernize packaging to use `pyproject.toml`
Also fixed up the NEWS file format (including incorrect line endings),
and moved the plugin to `.plugin` instead of the package/module root,
following our current standards & template for first-party plugins.
