--- setup.cfg
@@ -1,26 +1,0 @@
-[metadata]
-name = sopel-reddit
-version = 1.1.0
-description = A reddit plugin for Sopel IRC bots.
-author = dgw
-author_email = dgw@technobabbl.es
-url = https://github.com/sopel-irc/sopel-reddit
-license = Eiffel Forum License, version 2
-classifiers =
-    Intended Audience :: Developers
-    Intended Audience :: System Administrators
-    License :: Eiffel Forum License (EFL)
-    License :: OSI Approved :: Eiffel Forum License
-    Topic :: Communications :: Chat :: Internet Relay Chat
-
-[options]
-packages = find:
-zip_safe = false
-include_package_data = true
-install_requires =
-    sopel>=8.0.0.dev0,<9
-    praw>=4.0.0,<8.0.0
-
-[options.entry_points]
-sopel.plugins =
-    reddit = sopel_reddit
